backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.10.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
selenium==4.28.1
reportlab==4.2.5
scikit-learn==1.5.1
numpy==1.26.4
pandas==2.2.0
joblib==1.4.2
pymongo==4.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
PyJWT==2.9.0
email-validator==2.2.0
